LuceneFindRegisteredPathsServiceTest.kt in progress

+++ src/main/kotlin/org/lmarek/memory/refresher/document/RegisteredPath.kt
@@ -1,3 +1,3 @@
 package org.lmarek.memory.refresher.document
 
-data class RegisteredPath(val documentId: Long, val path: String)
+data class RegisteredPath(val documentId: Int, val path: String)
